se hicieron cambios y configuraciones en los testt
