+++ src/test/java/aut/testcreation/steps/GoogleSteps.java
@@ -1,7 +1,5 @@
 package aut.testcreation.steps;
 
-import aut.testcreation.pages.Homapage;
-import aut.testcreation.pages.HomePage;
 import aut.testplan.sprint.GoogleTestsAnaRunner;
 import io.cucumber.java8.En;
 import org.junit.jupiter.api.Assertions;
@@ -11,8 +9,6 @@ public class GoogleSteps extends GoogleTestsAnaRunner implements En{
 
     public GoogleSteps(){
 
-
-        Homapage googleHomePage;
 
         Before(2, GoogleTestsAnaRunner::setUp);
         After(GoogleTestsAnaRunner::tearDown);
