+++ src/test/java/aut/testcreation/testcases/TestHoteles.java
@@ -2,46 +2,62 @@ package aut.testcreation.testcases;
 
 import aut.testcreation.pages.HomePage;
 import aut.testcreation.pages.HotelesPage;
+import aut.testcreation.pages.HotelesResults;
 import framework.engine.selenium.DriverFactory;
 import framework.engine.selenium.SeleniumTestBase;
+import org.junit.After;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.openqa.selenium.By;
+
 
 
 public class TestHoteles extends SeleniumTestBase {
 
     private HomePage homepage;
     private HotelesPage hotelesPage;
-    @Test
-    public void rh001 () throws InterruptedException {
+
+    private HotelesResults hotelesResults;
+
+
+    @BeforeEach
+    public void preTest(){
         homepage = new HomePage(DriverFactory.getDriver());
         hotelesPage = new HotelesPage(homepage.getDriver());
+        hotelesResults = new HotelesResults(homepage.getDriver());
         homepage.navigateTo("https://www.rumbo.es");
         homepage.noCookies();
+    }
+
+    @AfterEach
+    public void posTest(){
+
+    }
+
+    @Test
+    public void rh001 () throws InterruptedException {
         homepage.irAHoteles();
         hotelesPage.completarBusqueda("Roma,Italia");
-        hotelesPage.Estrellas();
+        hotelesResults.Estrellas();
     }
 
     @Test
     public void rh002 () throws InterruptedException {
-        homepage = new HomePage(DriverFactory.getDriver());
-        hotelesPage = new HotelesPage(homepage.getDriver());
-        homepage.navigateTo("https://www.rumbo.es");
-        homepage.noCookies();
         homepage.irAHoteles();
         hotelesPage.completarBusqueda("Roma,Italia");
-        hotelesPage.motel();
+        hotelesResults.motel();
     }
 
     @Test
     public void rh003 () throws InterruptedException {
-        homepage = new HomePage(DriverFactory.getDriver());
-        hotelesPage = new HotelesPage(homepage.getDriver());
-        homepage.navigateTo("https://www.rumbo.es");
-        homepage.noCookies();
         homepage.irAHoteles();
-        hotelesPage.huespedes("Roma,Italia");
+        hotelesResults.huespedes("Roma,Italia");
+    }
+
+    @Test
+    public void rh004 () throws InterruptedException {
+        homepage.irAHoteles();
+        hotelesResults.huespedes("Roma,Italia");
     }
 
 }
